+++ spec/spec_helper.rb
@@ -1,8 +1,22 @@
 # frozen_string_literal: true
 
+ENV["BUNDLE_GEMFILE"] ||= File.expand_path("../Gemfile", __dir__)
+
+require "bundler/setup" if File.exist?(ENV["BUNDLE_GEMFILE"])
+
 require "stringio"
 require "sidekiq/job_logger"
 require "sidekiq/job_retry"
+
+begin
+  require "simplecov"
+  SimpleCov.start do
+    add_filter ["/spec/"]
+  end
+rescue LoadError
+end
+
+Bundler.require(:default, :test)
 
 require_relative "../lib/lumberjack_sidekiq"
 
@@ -29,4 +43,8 @@ class MySplatWorker
 
   def perform(arg1, *rest)
   end
+end
+
+if defined?(Lumberjack::Rails)
+  ActiveSupport::BroadcastLogger.include?(Lumberjack::Rails::BroadcastLoggerExtension)
 end